@@ -57,10 +57,10 @@
     "passport-azure-ad": "^3.0.9",
     "passport-jwt": "^3.0.1",
     "passport-local": "^1.0.0",
-    "react": "^15.6.2",
+    "react": "^16.4.0",
     "react-burger-menu": "^2.1.6",
     "react-cookie": "^0.4.7",
-    "react-dom": "~15.6.1",
+    "react-dom": "^16.0.0",
     "react-ga": "^1.4.1",
     "react-html-parser": "^2.0.2",
     "react-markdown": "^3.2.2",
@@ -108,7 +108,7 @@
     "mocha": "^3.2.0",
     "node-sass": "^4.5.3",
     "nodemon": "^1.11.0",
-    "prop-types": "^15.6.0",
+    "prop-types": "^15.6.1",
     "react-addons-test-utils": "^15.6.2",
     "sass-loader": "^6.0.6",
     "should": "^11.1.1",
