@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Field, reduxForm } from 'redux-form';
 import { getForgotPasswordToken } from '../../actions/auth';
+import PropTypes from 'prop-types';
 
 const form = reduxForm({
   form: 'forgotPassword',
@@ -9,7 +10,7 @@ const form = reduxForm({
 
 class ForgotPassword extends Component {
   static contextTypes = {
-    router: React.PropTypes.object,
+    router: PropTypes.object,
   }
 
   componentWillMount() {
