@@ -1,4 +1,4 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 // import marked from 'marked';
 import { CSSTransitionGroup } from 'react-transition-group'
